fix: Cash that rounds to nothing, purchases from a vehicle, and stale overrides
Distribute_Funds_To_Each_Teammate took an int and every refinery caller divides
before calling it.  The per-second trickle is FundsDistributedPerSec over the
base's operation time factor, which is 2 while the power plant is down -- so a
base whose trickle is under two credits a second paid its team nothing at all
unpowered.  The trickle paid while a harvester unloads is frame-seconds times
FundsPerSecond, a fraction at any normal frame rate, truncated to zero every
frame; nothing was paid and TotalFunds was never drawn down until the unload
timer expired and dumped the lot.  Increment_Money has always taken a float.

cPurchaseRequestEvent is a client request the server acts on, and the server
checked nothing about where the player was.  A player in a vehicle is not
standing at a terminal, whatever their client says.

Copy_Settings applied the human animation override and the loiter collection
when the new character names one and left the previous character's in place
when it does not.  Copy_Settings is how a purchase re-dresses an existing
soldier, so every character bought after an overridden one kept moving like the
one before it.

+++ Code/Combat/basecontroller.cpp
@@ -760,7 +760,7 @@ BaseControllerClass::Find_Building (BuildingConstants::BuildingType type)
 //
 ////////////////////////////////////////////////////////////////
 void
-BaseControllerClass::Distribute_Funds_To_Each_Teammate (int funds)
+BaseControllerClass::Distribute_Funds_To_Each_Teammate (float funds)
 {
 	//
 	//	Only let the server assign money
@@ -792,7 +792,7 @@ BaseControllerClass::Distribute_Funds_To_Each_Teammate (int funds)
 				//
 				//	Add the money to this player's total
 				//
-				player_data->Increment_Money (float(funds));
+				player_data->Increment_Money (funds);
 			}
 		}
 	}
